add downvoting abilities

+++ packages/proto-server/routes/rooms.js
@@ -2,7 +2,6 @@ const express = require("express");
 const router = express.Router();
 const roomClient = require("../lib/roomClient");
 
-/* Create a new User (register). */
 router.post("/", async (req, res, next) => {
   if (!req.body) return res.sendStatus(400);
   console.log(JSON.stringify(req.body));
@@ -38,18 +37,44 @@ router.get("/", async (req, res, next) => {
 });
 
 router.delete("/:roomName", async (req, res, next) => {
+  if (!req.params) return res.sendStatus(400);
+  console.log(JSON.stringify(req.params));
+
   const { roomName } = req.params;
-  let roomsDeleted;
 
   try {
-    roomsDeleted = await roomClient.deleteRoom(roomName);
-    // console.log("roomsDeleted: ", roomsDeleted);
+    await roomClient.deleteRoom(roomName);
     res.status(200);
     res.json({ status: "OK" });
   } catch (error) {
     console.log(error);
     res.status(404);
     res.send(err);
+  }
+});
+
+router.put("/:roomName", async (req, res, next) => {
+  if (!req.params) return res.sendStatus(400);
+  console.log(JSON.stringify(req.params));
+
+  const { roomName } = req.params;
+
+  try {
+    let newRoom;
+    // todo: refactor when followers are added, use findAndUpdateOrCreate
+    if (req.body.downvotes) {
+      if (req.body.downvotes !== 1) return res.sendStatus(400);
+      newRoom = await roomClient.downvoteRoom(roomName);
+    } else {
+      newRoom = await roomClient.createRoom(roomName);
+    }
+    // console.log("newRoom: ", newRoom);
+    res.status(201);
+    res.send(newRoom);
+  } catch (error) {
+    console.log(error);
+    res.status(400);
+    res.send(error);
   }
 });
